refactor: extract shared chart helpers into chartUtils.ts
