+++ apps/frontend-nextjs/app/components/dashboard/DashboardOverviewCard.tsx
@@ -3,43 +3,13 @@
 import { ITemaDashboardStats } from '@/constants/dashboard/dashboardConstants'
 import { BodyShort, Detail, Heading } from '@navikt/ds-react'
 import { Bar, BarChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts'
-
-const KRAV_COLORS = {
-  underArbeid: '#fa4d56',
-  ferdigVurdert: '#005d5d',
-}
-
-const SUKSESS_COLORS = {
-  underArbeid: '#1192e8',
-  oppfylt: '#005d5d',
-  ikkeOppfylt: '#fa4d56',
-  ikkeRelevant: '#9f1853',
-}
-
-interface IBarSegment {
-  name: string
-  value: number
-  color: string
-}
-
-const roundedPercentages = (values: number[]): number[] => {
-  const total = values.reduce((s, v) => s + v, 0)
-  if (total === 0) return values.map(() => 0)
-  const exact = values.map((v) => (v / total) * 100)
-  const floored = exact.map((v) => Math.floor(v))
-  const remainder = 100 - floored.reduce((s, v) => s + v, 0)
-  const diffs = exact.map((v, i) => ({ i, diff: v - floored[i] }))
-  diffs.sort((a, b) => b.diff - a.diff)
-  for (let j = 0; j < remainder; j++) {
-    floored[diffs[j].i] += 1
-  }
-  return floored
-}
-
-const formatPct = (pct: number, value: number): string => {
-  if (value > 0 && pct === 0) return '<1'
-  return `${pct}`
-}
+import {
+  IBarSegment,
+  KRAV_COLORS,
+  SUKSESS_COLORS,
+  formatPct,
+  roundedPercentages,
+} from './chartUtils'
 
 const OverviewStackedBar = ({
   data,
