+++ apps/frontend-nextjs/app/components/dashboard/TemaDashboardPage.tsx
@@ -16,43 +16,13 @@ import { DownloadIcon } from '@navikt/aksel-icons'
 import { BodyShort, Button, Heading, LocalAlert, Select, Tabs } from '@navikt/ds-react'
 import { useEffect, useState } from 'react'
 import { Bar, BarChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts'
-
-const KRAV_COLORS = {
-  underArbeid: '#fa4d56',
-  ferdigVurdert: '#005d5d',
-}
-
-const SUKSESS_COLORS = {
-  underArbeid: '#1192e8',
-  oppfylt: '#005d5d',
-  ikkeOppfylt: '#fa4d56',
-  ikkeRelevant: '#9f1853',
-}
-
-interface IBarSegment {
-  name: string
-  value: number
-  color: string
-}
-
-const formatPct = (pct: number, value: number): string => {
-  if (value > 0 && pct === 0) return '<1'
-  return `${pct}`
-}
-
-const roundedPercentages = (values: number[]): number[] => {
-  const total = values.reduce((s, v) => s + v, 0)
-  if (total === 0) return values.map(() => 0)
-  const exact = values.map((v) => (v / total) * 100)
-  const floored = exact.map((v) => Math.floor(v))
-  const remainder = 100 - floored.reduce((s, v) => s + v, 0)
-  const diffs = exact.map((v, i) => ({ i, diff: v - floored[i] }))
-  diffs.sort((a, b) => b.diff - a.diff)
-  for (let j = 0; j < remainder; j++) {
-    floored[diffs[j].i] += 1
-  }
-  return floored
-}
+import {
+  IBarSegment,
+  KRAV_COLORS,
+  SUKSESS_COLORS,
+  formatPct,
+  roundedPercentages,
+} from './chartUtils'
 
 const RechartsStackedBar = ({
   data,
